requirements.txt
Flask
psycopg2-binary # For PostgreSQL
# mysql-connector-python # Uncomment and use this for MySQL instead